Workaround for issue with CORS and Origin: null in vertx-web
The CORS handling mechanism of vertx-web should allow an HTTP header
with `Origin: null`, but currently it doesn't; see
https://github.com/vert-x3/vertx-web/issues/1933 for the details.

While that bug is fixed, this commit adds a workaround that enables
Kroki to handle that header value properly.

The workaround just defines a regexp that allows any value in that
header; vertx-web just matches the value using that regexp, and if it
matches the request continues its processing as normal.