docs(install.adoc): Split into multiple pages (#828)
The install.adoc page has grown and with the increasing number of sections
is becoming more difficult to comprehend and navigate.

The approach chosen is to split this content into multiple files while
preserving individual contributions the content of the new files,
and to then simplify the install pages so that each focuses
on its speciality.

docs: Overall changes
- Updates for asciidoc syntax, style, terms, DRY and consistency.
- Attributes: Use meaningful names (rename)
- Attributes: Remove unused
- Style: Use consistent Kroki language style
- Contributions: Preserve when moving sections to new files
-- Preserve line history for split out content using the method of:
-- https://devblogs.microsoft.com/oldnewthing/20190919-00/?p=102904
- Don't repeat yourself (DRY)
- Language: Containers provide diagrams
- Simplify
- Accuracy
- Use "To do this, you will need to ..." vs "To do this ..."
- Be specific about responsibilities
- Use proper names: Internet vs internet

docs(nav.adoc): Add new entries
- Add use-kubernetes.adoc
- Add use-docker-or-podman.adoc

docs(use-docker-or-podman.adoc):
- Preserve contributions in copy of install.adoc
- Rename attribute uri-localhost-8000 to uri-local-kroki-access

docs(use-kubernetes.adoc):
- Preserve contributions in copy of install.adoc
- Use {examplesdir}/k8s to refer to the individual definition yaml files
- Divide the guide into sections

docs(install.adoc):
- Copy with contributions to use-kubernetes.adoc
- Copy with contributions to use-docker-or-podman.adoc
- Remove redundant content
- Add prerequisites section
- Emphasize that the companion containers are optional.
- Simplify the Kroki container images section.
- Minimize distractions by pulling the headings for the kroki
container and companion containers up to level 1.
- Rework section "What you will learn"
- Rework section "Introduction"
- Rework section "Images"
- Rework section "Getting started"
Redo the Get Started section to be a self contained description
of what is required, with links to the tasks. It may repeat
some information but that is necessary to provide the reader
getting started with sufficient information.

Co-authored-by: Guillaume Grossetie <g.grossetie@gmail.com>