feat: add PGF/TikZ support
This change adds support for converting PGF/TikZ to JPEG, PDF, PNG, and
SVG.

The user input is lightly pre-processed.  If the user input contains a
\docmentclass command, the user input is passed to latex untouched.
Otherwise, a \documentclass command is derived from the output format
and prepended to the user input and then passed to latex.

For JPEG and PNG: Use latex to convert the user input to PDF.  Then
convert the PDF to JPEG or PNG using pdftocairo.

For PDF: Use latex to convert the user input to PDF.

For SVG: Use latex to convert the user input to DVI.  Then convert the
DVI to SVG using dvisvgm.

I also considered pdftocairo and pdf2svg to generate SVG.  Pdf2svg is no
longer maintained and the author recommends pdftocairo over pdf2svg.
Dvisvgm was choosen over pdftocairo because of its superior text
handling and smaller better optimized SVGs.  Dvisvgm has the ability to
render text using fonts while pdftocairo converts text to paths.  This
allows text selection in browsers and enables smaller file sizes.  In
one case, pdftocairo produced a 1.4MB SVG while dvisvgm produced a 60KB
SVG for the same image with little to no visible differences.