fix(tikz): fix bad render due to bad documentclass
The tikz svg output uses dvisvgm.  Dvisvgm requires the input latex to
specify the PGF dvisvgm driver by passing the dvisvgm option via the
documentclass.

Previous to this change, the kroki tikz output would generate a correct
documentclass if the user didn't provide one.  However, if the user
provided a documentclass the kroki tikz output would use it as is.  This
allowed users to specify an incorrect documentclass and then get
undesired output.

Fix by overwriting any user-provided documentclass.  The implication is
that users can no longer provide their own documentclass.  However, I
couldn't find any convincing usecases for this in the context of
diagrams.

This change implements option 1 from
https://github.com/yuzutech/kroki/issues/1870#issuecomment-2822066909.
The alternative of surgically fixing up the user-provided documentclass
instead of completely overwriting it was also considered but decided
against due to complexity.  This alternative could be added later if a
convincing usecase presents.

Fixes #1870